download zip for browsers without fs api
