+++ app/modules/fonts/font-context.ts
@@ -1,6 +1,7 @@
 import { raise } from "../common/error"
 import { omit } from "../common/omit"
 import { compactParams } from "../dom/compact-params"
+import { css } from "../dom/css"
 import { FontDict } from "./api.server"
 
 // query param format: ?fonts=Roboto:400,400italic;Roboto+Mono:400,400italic
@@ -148,4 +149,53 @@ export function makeFontContext(
     getClearSelectionLink,
     getDefaultPreviewVariant,
   }
+}
+
+export function getFontFiles(context: FontContext, origin: string) {
+  const selectedVariants = context.selectedFontList.flatMap((font) =>
+    font.variants.map((variant) => ({ font, variant })),
+  )
+
+  const fontFiles = selectedVariants.map(({ font, variant }) => {
+    const normalizedFamily = font.family.replaceAll(" ", "-")
+    const fileName = `${normalizedFamily}-${variant.weight}-${variant.style}.woff2`
+
+    const url = new URL(`/api/convert-to-woff2?url=${variant.url}`, origin)
+
+    const data = fetch(url.toString()).then((response) =>
+      response.ok
+        ? response.arrayBuffer()
+        : Promise.reject(new Error(response.statusText)),
+    )
+
+    return { font, variant, name: fileName, data }
+  })
+
+  const cssFile = {
+    name: "fonts.css",
+    data: Promise.all([
+      import("prettier/standalone.js"),
+      import("prettier/parser-postcss.js"),
+    ]).then(([prettier, postcssParser]) => {
+      const cssCode = fontFiles
+        .map(
+          ({ font, variant, name }) => css`
+            @font-face {
+              font-family: "${font.family}";
+              font-style: ${variant.style};
+              font-weight: ${variant.weight};
+              src: url("./${name}") format("woff2");
+            }
+          `,
+        )
+        .join("\n\n")
+
+      return prettier.default.format(cssCode, {
+        parser: "css",
+        plugins: [postcssParser.default],
+      })
+    }),
+  }
+
+  return [...fontFiles, cssFile]
 }
